move fill table to renderable utils
