+++ src/main/java/at/hannibal2/skyhanni/features/inventory/chocolatefactory/ChocolateShopPrice.kt
@@ -29,6 +29,7 @@ import at.hannibal2.skyhanni.utils.StringUtils.addStrikethorugh
 import at.hannibal2.skyhanni.utils.StringUtils.removeColor
 import at.hannibal2.skyhanni.utils.UtilsPatterns
 import at.hannibal2.skyhanni.utils.renderables.Renderable
+import at.hannibal2.skyhanni.utils.renderables.RenderableUtils
 import net.minecraft.item.ItemStack
 
 @SkyHanniModule
@@ -168,7 +169,7 @@ object ChocolateShopPrice {
             add(Renderable.string("§eChocolate available: §6${ChocolateAmount.CURRENT.formatted}"))
             // TODO add chocolate spend needed for next milestone
             add(Renderable.string("§eChocolate spent: §6${chocolateSpent.addSeparators()}"))
-            add(LorenzUtils.fillTable(table, padding = 5, itemScale = config.itemScale))
+            add(RenderableUtils.fillTable(table, padding = 5, itemScale = config.itemScale))
         }
     }
